requirements.txt
llama-index
llama-index-core
llama-index-llms-huggingface-api
llama-index-embeddings-huggingface
llama-index-readers-file
huggingface-hub
sentence-transformers
streamlit
python-dotenv
sqlparse
pandas